Update README.md
In this git hub repository the power Bi file is shared as google drive link. In this a large data set of airbnb which may have been modified for training purposes is analysed to draw insignts, findings using the PBI software, in which i HAVE shared screenshots for the various reports.